좌표 저장 수정

+++ src/main/java/hansung/ac/kr/fooding/domain/Coordinate.java
@@ -5,14 +5,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
+import javax.persistence.Column;
 import javax.persistence.Embeddable;
+import java.math.BigDecimal;
 
 @Embeddable
 @Getter
 @AllArgsConstructor @NoArgsConstructor
 public class Coordinate {
-    private float x;
-    private float y;
+
+    private double x;
+    private double y;
 
     @Override
     public String toString() {
